Add file in form input, and able to update the file!

+++ app/Http/Controllers/MahasiswaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
 
 class MahasiswaController extends Controller
 {
@@ -22,14 +23,24 @@ class MahasiswaController extends Controller
 
     public function create(Request $request)
     {
+        $data_akun = $request->session()->get("username");
+
+        if ($data_akun == null){
+            return redirect(route('login'));
+        }
+
+        $filename_full = $request->file('inputCV')->store('public/CV');
+        $filename = explode("/",$filename_full)[2];
+
         $request->validate([
             "nim"=>"required|min:12|max:12",
-            "name"=>"required"
+            "name"=>"required",
         ]);
 
         DB::table("mahasiswas")->insert([
             'nim' => $request->nim,
             'name' => $request->name,
+            'filename' => $filename,
         ]);
 
         return redirect(route("mahasiswa"));
@@ -37,18 +48,54 @@ class MahasiswaController extends Controller
 
     public function edit(Request $request, $id)
     {
-        DB::table("mahasiswas")->where('id', $id)->update([
-            'nim' => $request->nim,
-            'name' => $request->name,
+        $data_akun = $request->session()->get("username");
+
+        if ($data_akun == null){
+            return redirect(route('login'));
+        }
+
+        $request->validate([
+            "nim"=>"required|min:12|max:12",
+            "name"=>"required"
         ]);
+
+        if (empty($request['inputCV'])) {
+            DB::table("mahasiswas")->where('id', $id)->update([
+                'nim' => $request->nim,
+                'name' => $request->name,
+            ]);
+        }
+
+        else {
+            $filename_old = DB::table('mahasiswas')->where('id', $id)->first()->filename;
+            Storage::delete("public/CV/$filename_old");
+
+            $filename_full = $request->file('inputCV')->store('public/CV');
+            $filename = explode("/",$filename_full)[2];
+
+            DB::table("mahasiswas")->where('id', $id)->update([
+                'nim' => $request->nim,
+                'name' => $request->name,
+                'filename' => $filename,
+            ]);
+        }
 
         return redirect(route("mahasiswa"));
     }
 
 
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        $data_akun = $request->session()->get("username");
+
+        if ($data_akun == null){
+            return redirect(route('login'));
+        }
+
+        $filename = DB::table('mahasiswas')->where('id', $id)->first()->filename;
+        Storage::delete("public/CV/$filename");
         DB::table("mahasiswas")->where('id', $id)->delete();
+
 
         return redirect(route("mahasiswa"));
     }
